chore: moved CustomButton from screen mini folder to component folder
