+++ packages/screens/Mini/ConnectLedger/components/AllSet.tsx
@@ -1,9 +1,8 @@
 import React from "react";
 import { View } from "react-native";
 
-import { CustomButton } from "../../components/Button/CustomButton";
-
 import { BrandText } from "@/components/BrandText";
+import { CustomButton } from "@/components/buttons/CustomButton";
 import { SpacerColumn } from "@/components/spacer";
 import { neutral77 } from "@/utils/style/colors";
 import { fontMedium16, fontSemibold30 } from "@/utils/style/fonts";
