+++ packages/screens/Mini/Settings/components/CheckList.tsx
@@ -2,10 +2,10 @@ import React, { Fragment, useState } from "react";
 import { View } from "react-native";
 
 import { RedAlert } from "./RedAlert";
-import { CustomButton } from "../../components/Button/CustomButton";
 import Checkbox from "../../components/Checkbox/Checkbox";
 import MiniTextInput from "../../components/MiniTextInput";
 
+import { CustomButton } from "@/components/buttons/CustomButton";
 import { SpacerColumn } from "@/components/spacer";
 import { neutral22, neutralA3 } from "@/utils/style/colors";
 import { layout } from "@/utils/style/layout";
