+++ packages/screens/Mini/Wallet/SendingToriScreen.tsx
@@ -10,7 +10,6 @@ import chevronDownSVG from "../../../../assets/icons/chevron-down-white.svg";
 import teritoriSVG from "../../../../assets/icons/networks/teritori.svg";
 import questionSVG from "../../../../assets/icons/question-gray.svg";
 import teritoriCircleSVG from "../../../../assets/icons/tori-circle.svg";
-import { CustomButton } from "../components/Button/CustomButton";
 import MobileModal from "../components/MobileModal";
 import MiniTable from "../components/Table/MiniTable";
 import MiniTableRow from "../components/Table/MiniTableRow";
@@ -23,6 +22,7 @@ import { Dropdown } from "@/components/Dropdown";
 import { SVG } from "@/components/SVG";
 import { UserNameInline } from "@/components/UserNameInline";
 import { TertiaryBadge } from "@/components/badges/TertiaryBadge";
+import { CustomButton } from "@/components/buttons/CustomButton";
 import { SpacerColumn, SpacerRow } from "@/components/spacer";
 import { useFeedbacks } from "@/context/FeedbacksProvider";
 import { useGetAssets } from "@/hooks/wallet/useGetAssets";
