+++ packages/screens/Mini/components/BlurViewWrapper.tsx
@@ -3,11 +3,11 @@ import { Dispatch, PropsWithChildren, SetStateAction, useState } from "react";
 import { StyleProp, TextStyle, View, ViewStyle } from "react-native";
 import { SvgProps } from "react-native-svg";
 
-import { CustomButton } from "./Button/CustomButton";
 import eyeClosedSVG from "../../../../assets/icons/eye-closed.svg";
 
 import { BrandText } from "@/components/BrandText";
 import { SVG } from "@/components/SVG";
+import { CustomButton } from "@/components/buttons/CustomButton";
 import { CustomPressable } from "@/components/buttons/CustomPressable";
 import { SpacerColumn } from "@/components/spacer";
 import { useFeedbacks } from "@/context/FeedbacksProvider";
